+++ tests/test_config_flow_user.py
@@ -1,7 +1,6 @@
 """Test config flow for ThesslaGreen Modbus integration."""
 
 from types import SimpleNamespace
-from typing import Any
 from unittest.mock import AsyncMock, patch
 
 import pytest
@@ -235,85 +234,6 @@ async def test_form_user_success():
     assert result2["options"][CONF_DEEP_SCAN] is True
     assert result2["options"][CONF_MAX_REGISTERS_PER_REQUEST] == 5
 
-@pytest.mark.parametrize(
-    "registers,expected_note",
-    [
-        ("holding", "auto_detected_note_success"),
-        (None, "auto_detected_note_limited"),
-    ],
-)
-@pytest.mark.asyncio
-async def test_async_step_confirm_auto_detected_note(registers, expected_note):
-    """Test confirm step auto detected note for different register counts."""
-    flow = ConfigFlow()
-    flow.hass = SimpleNamespace(config=SimpleNamespace(language="en"))
-
-    flow._data = {
-        CONF_HOST: "192.168.1.100",
-        CONF_PORT: 502,
-        "slave_id": 10,
-        CONF_NAME: "My Device",
-    }
-    flow._device_info = {
-        "device_name": "ThesslaGreen AirPack",
-        "firmware": "1.0",
-        "serial_number": "123",
-    }
-    available_registers = {registers: [1]} if registers else {}
-    flow._scan_result = {
-        "available_registers": available_registers,
-        "capabilities": {},
-    }
-
-    translations = {
-        "auto_detected_note_success": "Auto-detection successful!",
-        "auto_detected_note_limited": "Limited auto-detection - some registers may be missing.",
-    }
-
-    with patch(
-        "homeassistant.helpers.translation.async_get_translations",
-        new=AsyncMock(return_value=translations),
-    ):
-        result = await flow.async_step_confirm()
-
-    assert result["type"] == "form"
-    assert result["step_id"] == "confirm"
-    assert result["description_placeholders"]["auto_detected_note"] == translations[expected_note]
-
-@pytest.mark.asyncio
-async def test_async_step_confirm_capabilities_only_bool():
-    """Ensure capabilities list includes only boolean fields."""
-    flow = ConfigFlow()
-    flow.hass = SimpleNamespace(config=SimpleNamespace(language="en"))
-
-    flow._data = {
-        CONF_HOST: "192.168.1.100",
-        CONF_PORT: 502,
-        "slave_id": 10,
-        CONF_NAME: "My Device",
-    }
-    flow._device_info = {}
-    flow._scan_result = {
-        "available_registers": {},
-        "capabilities": {
-            "temperature_sensors": {"t1"},
-            "expansion_module": True,
-            "temperature_sensors_count": 1,
-        },
-        "register_count": 1,
-    }
-
-    with patch(
-        "homeassistant.helpers.translation.async_get_translations",
-        new=AsyncMock(return_value={}),
-    ):
-        result = await flow.async_step_confirm()
-
-    placeholders = result["description_placeholders"]
-    assert "Expansion Module" in placeholders["capabilities_list"]
-    assert "Temperature Sensors" not in placeholders["capabilities_list"]
-    assert placeholders["capabilities_count"] == "1"
-
 @pytest.mark.asyncio
 async def test_unique_id_sanitized():
     """Ensure unique ID replaces colons in host with hyphens."""
@@ -350,103 +270,4 @@ async def test_unique_id_sanitized():
 
     mock_set_unique_id.assert_called_once_with("fe80--1:502:10")
 
-@pytest.mark.asyncio
-async def test_duplicate_configuration_aborts():
-    """Test configuring same host/port/slave twice aborts the flow."""
-    flow = ConfigFlow()
-    flow.hass = None
-
-@pytest.mark.asyncio
-async def test_confirm_step_aborts_on_existing_entry():
-    """Ensure confirming a second flow aborts if unique ID already configured."""
-
-    flow = ConfigFlow()
-    flow.hass = None
-
-    user_input = dict(DEFAULT_USER_INPUT)
-
-    validation_result = {
-        "title": "Device",
-        "device_info": {},
-        "scan_result": {},
-    }
-
-    # First pass through user step to store data
-    validation_result = {
-        "title": "ThesslaGreen 192.168.1.100",
-        "device_info": {},
-        "scan_result": {},
-    }
-
-    class AbortFlow(Exception):
-        def __init__(self, reason: str) -> None:
-            self.reason = reason
-
-    entries: set[str] = set()
-
-    async def async_set_unique_id(self, unique_id: str, **_: Any) -> None:
-        self._unique_id = unique_id
-
-    def abort_if_unique_id_configured(self) -> None:
-        if getattr(self, "_unique_id", None) in entries:
-            raise AbortFlow("already_configured")
-
-    with (
-        patch(
-            "custom_components.thessla_green_modbus.config_flow.validate_input",
-            return_value=validation_result,
-        ),
-        patch(
-            "custom_components.thessla_green_modbus.config_flow.ConfigFlow.async_set_unique_id",
-        ),
-        patch(
-            "custom_components.thessla_green_modbus.config_flow.ConfigFlow."
-            "_abort_if_unique_id_configured",
-        ),
-    ):
-        await flow.async_step_user(user_input)
-
-    # Attempt to confirm after a duplicate has been configured elsewhere
-    with (
-        patch("custom_components.thessla_green_modbus.config_flow.ConfigFlow.async_set_unique_id"),
-        patch(
-            "custom_components.thessla_green_modbus.config_flow.ConfigFlow."
-            "_abort_if_unique_id_configured",
-            side_effect=RuntimeError("already_configured"),
-        ),
-        pytest.raises(RuntimeError),
-    ):
-        await flow.async_step_confirm({})
-
-    with (
-        patch(
-            "custom_components.thessla_green_modbus.config_flow.validate_input",
-            return_value=validation_result,
-        ),
-        patch(
-            "homeassistant.helpers.translation.async_get_translations",
-            new=AsyncMock(return_value={}),
-        ),
-        patch.object(ConfigFlow, "async_set_unique_id", async_set_unique_id),
-        patch.object(
-            ConfigFlow,
-            "_abort_if_unique_id_configured",
-            abort_if_unique_id_configured,
-        ),
-    ):
-        flow1 = ConfigFlow()
-        flow1.hass = SimpleNamespace(config=SimpleNamespace(language="en"))
-        flow2 = ConfigFlow()
-        flow2.hass = SimpleNamespace(config=SimpleNamespace(language="en"))
-
-        await flow1.async_step_user(user_input)
-        await flow2.async_step_user(user_input)
-
-        result1 = await flow1.async_step_confirm({})
-        assert result1["type"] == "create_entry"
-        entries.add(f"{user_input[CONF_HOST]}:{user_input[CONF_PORT]}:{user_input['slave_id']}")
-
-        with pytest.raises(AbortFlow) as err:
-            await flow2.async_step_confirm({})
-        assert err.value.reason == "already_configured"
